task2/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
pydantic==2.5.3
google-generativeai==0.3.2
python-multipart==0.0.6
python-dotenv==1.0.0

# For PostgreSQL support (if deploying to Heroku/Render with PostgreSQL)
psycopg2-binary==2.9.9

# For CORS
python-jose==3.3.0